requirements.txt
alembic==1.6.5
allure-pytest==2.8.6
allure-python-commons==2.8.6
APScheduler==3.7.0
asttokens==2.4.1
atomicwrites==1.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
colorlog==5.0.1
ConfigArgParse==1.5.1
crontab==0.23.0
Crypto==1.4.1
cryptography==43.0.1
decorator==5.0.9
defusedxml==0.7.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
Faker==8.10.1
fastjsonschema==2.20.0
Flask==1.1.2
Flask-APScheduler==1.12.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
gevent==21.1.2
geventhttpclient==1.4.5
gitdb==4.0.9
GitPython==3.1.24
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==8.5.0
iniconfig==2.0.0
ipython==8.12.3
itsdangerous==2.0.1
jedi==0.19.1
Jinja2==3.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-pygments==0.3.0
locust==1.6.0
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.5.0
msgpack==1.0.2
Naked==0.1.32
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pactverify==2.0.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.3.0
prompt-toolkit==3.0.48
psutil==5.8.0
pure-eval==0.2.3
py==1.11.0
pycparser==2.20
pycryptodome==3.21.0
pygments==2.18.0
PyMySQL==1.0.2
pytest==8.0.2
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-repeat==0.9.3
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2021.1
pywin32==308
PyYAML==5.4.1
pyzmq==22.1.0
redis==3.4.1
referencing==0.35.1
requests==2.26.0
requests-toolbelt==0.9.1
rpds-py==0.20.0
shellescape==3.8.1
six==1.16.0
smmap==5.0.0
soupsieve==2.6
SQLAlchemy==1.4.0
stack-data==0.6.3
text-unidecode==1.3
tinycss2==1.3.0
tomli==2.0.2
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.0.1
tzlocal==2.1
urllib3==1.26.6
validators==0.18.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.0.1
WTForms==2.3.3
xlrd==2.0.1
XMind==1.2.0
yarg==0.1.9
zipp==3.20.2
zope.event==4.5.0
zope.interface==5.4.0
